requirements.txt
opencv_python==4.5.1.48
nltk==3.5
numpy==1.20.1
scipy==1.6.0
lxml==4.5.2
editdistance==0.5.3
matplotlib==3.3.4
Send2Trash==1.5.0
pytesseract==0.3.4
selenium==3.141.0
imageio==2.9.0
pandas==1.2.1
networkx==2.4
gensim==3.8.3
Appium_Python_Client==1.1.0
Pillow==8.2.0
pycurl==7.43.0.6
pypng==0.0.20
scikit_learn==0.24.1
